+++ build_runner/lib/src/entrypoint/options.dart
@@ -9,6 +9,7 @@ import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:build_config/build_config.dart';
 import 'package:io/io.dart';
+import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 import 'package:shelf/shelf_io.dart';
 
@@ -17,6 +18,7 @@ import 'package:build_runner/build_runner.dart';
 const _assumeTty = 'assume-tty';
 const _define = 'define';
 const _deleteFilesByDefault = 'delete-conflicting-outputs';
+const _logRequests = 'log-requests';
 const _lowResourcesMode = 'low-resources-mode';
 const _failOnSevere = 'fail-on-severe';
 const _hostname = 'hostname';
@@ -113,10 +115,12 @@ class _SharedOptions {
 /// Options specific to the [_ServeCommand].
 class _ServeOptions extends _SharedOptions {
   final String hostName;
+  final bool logRequests;
   final List<_ServeTarget> serveTargets;
 
   _ServeOptions._({
     @required this.hostName,
+    @required this.logRequests,
     @required this.serveTargets,
     @required bool assumeTty,
     @required bool deleteFilesByDefault,
@@ -158,6 +162,7 @@ class _ServeOptions extends _SharedOptions {
     }
     return new _ServeOptions._(
       hostName: argResults[_hostname] as String,
+      logRequests: argResults[_logRequests] as bool,
       serveTargets: serveTargets,
       assumeTty: argResults[_assumeTty] as bool,
       deleteFilesByDefault: argResults[_deleteFilesByDefault] as bool,
@@ -311,7 +316,11 @@ class _ServeCommand extends _WatchCommand {
   _ServeCommand() {
     argParser
       ..addOption(_hostname,
-          help: 'Specify the hostname to serve on', defaultsTo: 'localhost');
+          help: 'Specify the hostname to serve on', defaultsTo: 'localhost')
+      ..addFlag(_logRequests,
+          defaultsTo: false,
+          negatable: false,
+          help: 'Enables logging for each request to the server.');
   }
 
   @override
@@ -343,11 +352,22 @@ class _ServeCommand extends _WatchCommand {
         trackPerformance: options.trackPerformance,
         verbose: options.verbose,
         builderConfigOverrides: options.builderConfigOverrides);
-    var servers = await Future.wait(options.serveTargets.map((target) =>
-        serve(handler.handlerFor(target.dir), options.hostName, target.port)));
+    var servers = await Future.wait(options.serveTargets.map((target) => serve(
+        handler.handlerFor(target.dir, logRequests: options.logRequests),
+        options.hostName,
+        target.port)));
     await handler.currentBuild;
-    for (var target in options.serveTargets) {
-      stdout.writeln('Serving `${target.dir}` on port ${target.port}');
+    // Warn if in serve mode with no servers.
+    if (options.serveTargets.isEmpty) {
+      var logger = new Logger('Serve');
+      logger.warning(
+          'Found no known web directories to serve, but running in `serve` '
+          'mode. You may expliclity provide a directory to serve with trailing '
+          'args in <dir>[:<port>] format.');
+    } else {
+      for (var target in options.serveTargets) {
+        stdout.writeln('Serving `${target.dir}` on port ${target.port}');
+      }
     }
     await handler.buildResults.drain();
     await Future.wait(servers.map((server) => server.close()));